Code enhancement

+++ src/libserver/data/file/FileDataSource.cpp
@@ -601,18 +601,18 @@ void server::FileDataSource::RetrieveHorse(data::Uid uid, data::Horse& horse)
   if (json.contains("breeding"))
   {
     horse.breeding.timesBreeded = json["breeding"].value("timesBreeded", uint32_t{0});
-    horse.breeding.breedingCombo = json["breeding"].value("breedingCombo", uint8_t{0});
+    horse.breeding.breedingCombo = json["breeding"].value("breedingCombo", uint32_t{0});
   }
   else
   {
     // Legacy flat structure
     horse.breeding.timesBreeded = json.value("timesBreeded", uint32_t{0});
-    horse.breeding.breedingCombo = json.value("breedingCombo", uint8_t{0});
+    horse.breeding.breedingCombo = json.value("breedingCombo", uint32_t{0});
   }
 
-  horse.horseType = json.value("horseType", uint8_t{0});
-  horse.tendency = json.value("tendency", uint8_t{0});
-  horse.spirit = json.value("spirit", uint8_t{0});
+  horse.horseType = json.value("horseType", uint32_t{0});
+  horse.tendency = json.value("tendency", uint32_t{0});
+  horse.spirit = json.value("spirit", uint32_t{0});
   horse.fatigue = json.value("fatigue", uint16_t{0});
 
   const auto& breedingJson = json["breeding"];
@@ -660,7 +660,7 @@ void server::FileDataSource::RetrieveHorse(data::Uid uid, data::Horse& horse)
     horse.ancestors = std::vector<data::Uid>{};
   }
   
-  horse.lineage = json.value("lineage", uint8_t{1});
+  horse.lineage = json.value("lineage", uint32_t{1});
 }
 
 void server::FileDataSource::StoreHorse(data::Uid uid, const data::Horse& horse)
